paypre_mailServices/paypre-api-event/requirements.txt
aio-pika==9.1.3
aioodbc==0.4.0
aiormq==6.7.6
anyio==3.6.2
arabic-reshaper==3.0.0
asn1crypto==1.5.1
async-timeout==4.0.2
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
cssselect2==0.7.0
dnspython==2.3.0
fastapi==0.95.1
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
joblib==1.2.0
lxml==4.9.2
multidict==6.0.4
oscrypto==1.3.0
pamqp==3.2.1
passlib==1.7.4
Pillow==9.5.0
pybars3==0.9.7
pycparser==2.21
pydantic==1.10.7
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
PyMeta3==0.5.1
pymongo==4.3.3
pyodbc==4.0.39
pypdf==3.11.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
PyYAML==6.0
qrcode==7.4.2
redis==4.6.0
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
uvicorn==0.21.1
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.2
xhtml2pdf==0.2.11
yarl==1.9.2
